+++ src/main_opm.py
@@ -119,40 +119,4 @@ if __name__ == "__main__":
   causalName = 'causal_keywords_entity_ruler'
   matcher.addEntityPattern(causalName, patternsCausal)
 
-  # ##Issue with simple and phrase matcher, if there are duplicate names, callback functions
-  # ##can not be used, in which case, we can not directly extend doc.ents, which will raise the
-  # ## error:  Unable to set entity information for token xx which is included in more than
-  # ## one span in entities, blocked, missing or outside. (https://github.com/explosion/spaCy/discussions/9993)
-  # ## In this case, we need to use "entity_ruler" to identify the matches
-  # # simple match
-  # name = 'ssc_match'
-  # rules = [{"LOWER":"pump"}, {"POS":"NOUN"}]
-  # matcher.addPattern(name, rules, callback=None)
-  # # phrase match
-  # name = 'ssc_phrase_match'
-  # phraseList = ['pump 1A', 'pump bearings', 'Pump']
-  # matcher.addPhrase(name, phraseList, callback=None)
-  # # depency match
-  # name = 'ssc_dependency_match'
-  # dependencyList = [
-  #   {
-  #     "RIGHT_ID":"anchor_found",
-  #     "RIGHT_ATTRS":{"ORTH":"found"}
-  #   },
-  #   {
-  #     "LEFT_ID":"anchor_found",
-  #     "REL_OP":">",
-  #     "RIGHT_ID":"subject",
-  #     "RIGHT_ATTRS":{"DEP":"nsubjpass"}
-  #   },
-  #   {
-  #     "LEFT_ID":"anchor_found",
-  #     "REL_OP":">",
-  #     "RIGHT_ID":"object",
-  #     "RIGHT_ATTRS":{"DEP":"oprd"}
-  #   }
-  # ]
-  # matcher.addDependency(name, dependencyList, callback=None)
-
-
   matcher(doc)
