+++ src/nlp/nlp_utils.py
@@ -8,6 +8,9 @@ Created on March, 2022
 import pandas as pd
 import spacy
 from spacy.tokens import Span
+# filter_spans is used to resolve the overlap issue in entities
+# It gives primacy to longer spans (entities)
+from spacy.util import filter_spans
 import logging
 
 import networkx as nx
@@ -101,3 +104,21 @@ def plotDAG(edges, colors='k'):
   ax = plot.gca()
   plt.axis("off")
   plot.show()
+
+
+###############
+# methods can be used for callback in "add" method
+###############
+def extendEnt(matcher, doc, i, matches):
+  """
+    Extend the doc's entity
+    @ In, matcher, spacy.Matcher, the spacy matcher instance
+    @ In, doc, spacy.tokens.doc.Doc, the processed document using nlp pipelines
+    @ In, i, int, index of the current match (matches[i])
+    @ In, matches, List[Tuple[int, int, int]], a list of (match_id, start, end) tuples, describing
+      the matches. A match tuple describes a span doc[start:end]
+  """
+  id, start, end = matches[i]
+  ent = Span(doc, start, end, label=id)
+  logger.debug(ent.label_ + ' ' + ent.text)
+  doc.ents = filter_spans(list(doc.ents) +[ent])
